refactor - Direction을 Enum으로 변경
